requirements.txt
aniso8601==8.0.0
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dnspython==1.16.0
Flask==1.1.1
Flask-RESTful==0.3.8
idna==2.5
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
requests2==2.16.0
selenium==3.141.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.21.1
Werkzeug==1.0.0
